removed redundant segments file
