+++ .firebase/hosting.ZnJvbnRlbmQvYnVpbGQ.cache
@@ -1,15 +1,14 @@
-asset-manifest.json,1718084356752,f2e692e8c4cebfa3908b1a80c8420f747ef82df3e30235e74e181582de7c24d0
-index.html,1718084356752,ce1babe9ea1b63838e66b6ca4fd2ec65f698bcbdac7d4122c4b4a186ef21bf8e
-robots.txt,1718084350518,391d14b3c2f8c9143a27a28c7399585142228d4d1bdbe2c87ac946de411fa9a2
-error.png,1718084350517,58fe103aecc43d648eedfa98dec343acc2aae211244bb33000dded64ee653149
-manifest.json,1718084350518,6aaf3d87dd9bb1e96f9c691c50d0ed378ed98a120a76dea78291281834cd36e9
-tv-television-svgrepo-com.svg,1718084350519,1af4174d84ffc60efe8ebdd892d4aecb1359fab9753fe2623374d751055c9452
-static/css/main.332df4c8.css,1718084356765,ded754ff69d5f667685db4bdb96cf2067b7d7bbf3c83c531c6d7a5dd131e6ecb
-static/css/main.332df4c8.css.map,1718084356765,45e32626f5151c9af160fbf6299a7c49a1ee908d423018ce74941721a8b9a9d5
-static/js/main.ddda4fb0.js.LICENSE.txt,1718084356764,d840d3fce21e0ad5aac21bd529b293c7fe5858973e070ed41ffd9566d4df62c3
-tvicon.png,1718084350520,1ae40736f6d522230aa634450f95263f12852d2c67d43451198d69ea536e60f5
-Gargoyle.jpg,1718084350517,39cb8a6977c875f34e43b8436fc764204cd7b33120df6e604078057046531536
-tv_transparent.png,1718084350519,edc9c13dfa12b7106d573d012b03a1086de97e64e1b342aeacf4091fc6735903
-title-screenshot.png,1718084350518,c0744e01d36d23d98fdbd172476e8895d93c79101495b9aa0b9b08ea5b81a448
-static/js/main.ddda4fb0.js,1718084356765,944f323ad7db470ad9e1b8d843f46b75941719371e9cc907875072e494365520
-static/js/main.ddda4fb0.js.map,1718084356765,413d2954dd0695417419d6a8b65d82172504824ba7dfff630ff5a0a724820b6a
+index.html,1718150338226,f2bd0ba4a650b4dd3c49db8997b6e8bd99cbb4da35abbfbd85797e060d717719
+asset-manifest.json,1718150338226,46448a4b8aeb468c609e02bb0d981adec1acdd5aa3510e7a8d3152009ff7792f
+error.png,1718150331938,58fe103aecc43d648eedfa98dec343acc2aae211244bb33000dded64ee653149
+robots.txt,1718150331938,391d14b3c2f8c9143a27a28c7399585142228d4d1bdbe2c87ac946de411fa9a2
+manifest.json,1718150331938,6aaf3d87dd9bb1e96f9c691c50d0ed378ed98a120a76dea78291281834cd36e9
+static/css/main.52261eb2.css,1718150338238,78c22c3dad62298346597ffdeb0d9a0d47972d0e659d1a0a0d7b2af3a91d02b0
+static/js/main.453565fa.js.LICENSE.txt,1718150338238,d840d3fce21e0ad5aac21bd529b293c7fe5858973e070ed41ffd9566d4df62c3
+tvicon.png,1718150331940,1ae40736f6d522230aa634450f95263f12852d2c67d43451198d69ea536e60f5
+static/css/main.52261eb2.css.map,1718150338237,52c3573b6d50f06c9b32793f6fbceaf57ecd57c6aab19ebf47f23684187cbabb
+Gargoyle.jpg,1718150331937,39cb8a6977c875f34e43b8436fc764204cd7b33120df6e604078057046531536
+tv_transparent.png,1718150331939,edc9c13dfa12b7106d573d012b03a1086de97e64e1b342aeacf4091fc6735903
+title-screenshot.png,1718150331939,c0744e01d36d23d98fdbd172476e8895d93c79101495b9aa0b9b08ea5b81a448
+static/js/main.453565fa.js,1718150338238,f9490658f2f96bf70cf3d91aa58147177904f0f9ac06705268237055cf4016c6
+static/js/main.453565fa.js.map,1718150338237,6a3ec5bbef473251c0ad28b3c4c7a67b2baa59363bf61698ad64f5826b1a8a44
